Styling

+++ javascript/src/eventListeners.js
@@ -12,17 +12,35 @@ hideNoteButton.addEventListener("click", function(event) {
 
 let clearNoteButton = document.getElementById('clear-notes');
 clearNoteButton.addEventListener("click", function(event) {
-  window.localStorage.clear();
-  document.getElementById("list").innerHTML = "";
+  confirmClear();
 })
 
 let deleteNoteButton = document.getElementById('delete-note');
 deleteNoteButton.addEventListener("click", function(event) {
-  let text = noteBody.textContent;
-  let index = storedNoteArray.indexOf(text);
-  storedNoteArray.splice(index, 1);
-  window.localStorage.setItem('listOfNotes', JSON.stringify(storedNoteArray))
-  document.getElementById("list").innerHTML = "";
-  hideNote()
-  updateNoteList()
+  confirmDelete();
 })
+
+function confirmClear() {
+  let response = confirm("Are you sure? This action cannot be undone!");
+  if (response === true) {
+    window.localStorage.clear();
+    clearList();
+  }
+}
+
+function confirmDelete() {
+  let response = confirm("Are you sure? This action cannot be undone!")
+  if (response === true) {
+    let text = noteBody.textContent;
+    let index = storedNoteArray.indexOf(text);
+    storedNoteArray.splice(index, 1);
+    window.localStorage.setItem('listOfNotes', JSON.stringify(storedNoteArray));
+    clearList();
+    hideNote();
+    updateNoteList();
+  }
+}
+
+function clearList() {
+  document.getElementById("list").innerHTML = "";
+}
